backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data validation
pydantic==2.10.0

# Testing
pytest==8.3.0
httpx==0.28.0
pytest-asyncio==0.24.0
pytest-cov==4.1.0

# OpenAPI contract testing
openapi-core==0.18.2

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
